Added DynamicClock::setFreq(num, denom)
Added a method to set the frequency as a ratio. Compared to the existing
setFreq() method which accepts an integer parameter, using this new method can
result in less rounding errors. This new method will be used in the next patch.

At the same time simplified the corresponding calculation in the Clock class.
This calculation was still prepared to deal with 96-bit intermediate results
(not needed anymore). This calculation was correct and it happened at compile
time (so speed is not an issue). But it's nicer to use the same formula in both
Clock (compile-time) and DynamicClock (run-time).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12254 7512d703-071c-0410-b5c2-d2bdd895b7d1