Made EmuTimer class more 'dynamic'
Before this patch, the EmuTimer class was a template that could be instantiated
to get a specific timer (e.g. YM2151 and Y8950 used different instances of this
template). This template approach results in the fastest code (all details of
the timer are known at compile-time), but it is also less flexible. For example
it's not possible anymore to at run-time switch between two different timer
configurations. But this is exactly what we needed.

The OPL3 and OPL4 (YMF262, YMF278) are mostly compatible, but there are small
differences. One such difference is the exact timing of the two provided
timers. Because both chips use a different input clock, the timers also run at
a slightly different rate.

This patch de-templatizes the EmuTimer class (all template parameters are now
run-time variables). This is slower, but this also isn't performance critical
(I didn't check but I believe EmuTimer was originally implemented with
templates because at the time the DynamicClock class didn't exist yet). But it
does add the flexibility to at-run-time choose between the OPL3 or the OPL4
timings.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12255 7512d703-071c-0410-b5c2-d2bdd895b7d1