Fixed (small) error in OPL4 (FM-part) sample rate
The OPL3 (YMF262) (and also MSX-AUDIO (Y8950)) have a sample frequency of
  (4 * 3579545) / (8 * 36) = 49716Hz
But the FM-part of the OPL4 (YMF278) has a sample frequency of
  33868800 / (19 * 36)     = 49516Hz
This patch fixes this differece (thanks n_n for reminding me).

The difference between the two is small (less than 0.5%) but still it's strange
that two supposed to be compatible chips have this difference. Also if you look
at the datasheet, the recommended FNUM values to generate notes have a similar
difference (so this confirms the input clock difference).

The reason for this difference is probably the OPL4 wave-part. That part has
an output rate of 44100Hz and 33868800 is an integer multiple of 44100 (and the
0.5% difference is likely small enough to not really notice the difference).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12256 7512d703-071c-0410-b5c2-d2bdd895b7d1