Remove ancient (r1110) hack to fix 64k ROMs on ASCII16 mapper.
It is not true that all 64k ROMs are of ASCII16 mapper type. An example
is the Auf Wiedersehen Monty ROM version (a hack, but works), which is
of ASCII8 type.

At the time the hack was introduced, 64k ROMs didn't work at all
(example: The Apeman Strikes Again). Without the hack, they do work now,
even though they are not in the softwaredb.

I tested this with about 50 64k ROMs. There was one regression
(spirits.rom, 8afd7312b10626fb7e1c1692b2c46d29b2ec06a6, which is now
detected as a Konami type).

All in all, this made me decide to remove the hack completely. An
alternative is to add it later on in the detection code as a fallback.
But so far, I have not seen any indication that this is needed. It
definitely doesn't fix the regression. Weird things like the regression
should probably be solved by softwaredb.xml entries.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12258 7512d703-071c-0410-b5c2-d2bdd895b7d1