Optimize calculation of 'step'
Idea copied from n_n's changes in MAME. The code contained an expression
'x << y' with 'y' possibly a negative value (in which case we should shift in
the other direction). So the original code required a test followed by two
expression with reversed shift. n_n optimized this to '(x << (N + y)) >> N'
(with 'y + N >= 0').

I did a 2nd optimization: 4-bit sign-extension was implemented as
  if (x & 8) x |= -8;
But this is faster
  (x ^ 8) - 8;
And it can often be (partly) folded into the surrounding expression.

The code uses the YMF278Slot::step member variable to cache some calculation
(at any point this variable can be recalculated from the 'OCT' and 'FN' data
members). Though this invariant was violated in the YMF278Slot::reset() method
and because of this is was required to recalculate 'step' in the
YMF278Impl::keyOnHelper() method. This patch fixes the reset() method and
simplifies keyOnHelper(). Now 'step' is only recalculated when either 'OCT' or
'FN' changes. Because 'step' is now fully redundant we also don't serialize it
anymore.

Added some TODO's about possible future optimizations (store 'OCT' and 'FN'
member variables in a different way).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12263 7512d703-071c-0410-b5c2-d2bdd895b7d1