Improved emulation of the BUSY status bit
The BUSY bit in the status register was not yet set for wave register select
and for FM register select and register write. For wave registers 3-5 the
timing was also off.

The emulation of the BUSY (and LD) status bits was moved from the YMF278 to the
MSXMoonSound class. The reason is that in the YMF278 class the FM part of the
chip is not visible nor writes to the wave register select register. The
MSXMoonSound class does have all the required information.

This patch is triggered by n_n's changes in ymf278b.c in the MAME code. It also
improves upon n_n's work (e.g. timing for wave registers 3-6). The improvements
are based on the datasheet and also verified on real HW.

These are the type of tests I performed on a real YMF278 (MoonSound cartridge):
	; test program for wave-register-select
	ld c,#c4      ; status register
	ld a,4        ; wave register 4
	out (#7e),a   ; select wave register
	in e,(c)      ; read
	in d,(c)      ;   status
	in l,(c)      ;      register
	in h,(c)      ;         4 times
	out (#7f),a   ; actually write register
So we do an action (select wave register in this example) and directly after
read the status register 4 times as fast as possible. Below is a table with the
results:

  Action             | BUSY-bit
                     | R800  Z80
---------------------+------------
wave register select | #000  0000
wave register write  | #000  0000
wave R#3-R#6  write  | 0000  0000
wave register read   | 0000  0000
FM   register select | *000  0000
FM   register write  | *000  0000
FM   register read   | 0000  0000
  # -> mostly '1', but occasionally '0'    (probably because of R800 refresh)
  * -> mostly '0', but occasionally '1'

Conclusion:
On Z80 the BUSY bit always reads as '0' (Z80 is too slow to see the '1').
On R800 the BUSY bit is '1' briefly after a wave register select or write and
_very_ briefly after a FM register select or write. Writing to wave registers
R#3-R#6 is faster than to the other wave registers. According to the datasheet
the BUSY bit will also be '1' after reading from wave register R#6, but even in
R800 mode this time is too short to be able to measure.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12265 7512d703-071c-0410-b5c2-d2bdd895b7d1