Implemented memory based I/O for the MSX RS-232C.
This is used by later RS-232C interfaces, like the Sony HBI-232 and the
one in the Sony HB-G900P and Sony HB-G900AP. It has a mechanism to also
enable port-based access "to be compatible with MSX1". The information
used to implement it comes from the Sony HB-G900P and HB-G900AP service
manual. Without this, the RS-232C ROM doesn't initialize properly.

To enable the memory based access, you need to specify a tag in the
MSX-RS232 element in the hardwareconfig.xml:
<memorybasedio>true</memorybasedio>.

Thanks to BiFiMSX for supporting the implementation work.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12290 7512d703-071c-0410-b5c2-d2bdd895b7d1