Fixed 'watchpoints don't break immediately' bug.
To reproduce simply execute:
  debug set_watchpoint write_io 0xa1
in MSX-BASIC on a NMS8250, this should stop at address 0x1217 (directly after
the 'OUT (#a1),A' instruction. For some strange reason (possibly just (bad)
luck) I couldn't reproduce this when using a 'devel' build. With either an
'opt' or even a 'debug' build I could trigger this bug.

The problem was that the 'debug break' command didn't force the CPU emulation
to exit the fast-emulation loop. 'Typically' this also isn't required: at the
time commands are accepted, the CPU emulation loop isn't running. Or in case of
breakpoints or debug conditions, the fast cpu loop is never entered. Only for
watchpoints this can happen.

Fixed by letting the 'debug break' command call CPU::exitCPULoopSync().

Thanks to BiFiMSX for reporting the bug, and insisting there really was a bug
even though I couldn't reproduce it at first.



git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12313 7512d703-071c-0410-b5c2-d2bdd895b7d1