Fixed bug in 'reverse loadreplay'
I can trigger the problem like this:
 - start and pause openmsx (I paused soon (before the bootlogo) I don't know
   if that matters)
 - in the console excute the command 'reverse loadreplay <any-replay file>'
 - execute the same command again (all the time openmsx remains in pause mode)
=> this triggered an assert for me

The bug was introduced in revision 12223. I that commit we added an
optimization for the 'reverse goto' command: if we're taking a (small) forward
step in time it might be better to continue from the current position than to
first load an (older) snapshot. Though this optimization only works when we're
moving in the same time-line. Internally 'reverse loadreplay' also uses the
'goto' function, but this moves to the time-line from the replay file.

Fixed by disabling the above optimization for 'loadreplay'.


git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12315 7512d703-071c-0410-b5c2-d2bdd895b7d1