Fixed slowness when skipping forward in replay file.
To reproducer: load a long replay (the effect is magnified on long replay files
with only the initial snapshot). Next skip forward e.g. by clicking in the
reversebar. After some (reasonable amount of) time the frame of the target time
is shown. But then it still takes an even longer time before openMSX becomes
responsive again.

What is happening is that during the time openmsx is unresponsive, it is
processing all the piled-up FINISH_FRAME events (so it is drawing the same
frame hundreds of times). This bug is another fallout of revision 12223.
Normally during fast-forward we skip forward till a few frames before the
target time without frame rendering and forward the last few frames with frame
rendering enabled. Though the mechanism that disabled frame rendering didn't
work in combination with the optimization of revision 12223.

Fixed by making the fast-forward state explicit.


git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12316 7512d703-071c-0410-b5c2-d2bdd895b7d1