Implemented basic support for V9990 set adjust
I didn't yet test when exactly changes to the adjust register take effect on a
real V9990 (at the start of a line or on the start of a frame or immediately).
For comparison on a V99x8 vertical set adjust changes take effect on the next
frame and horizontal set adjust changes take effect the next line.

BiFiMSX asked me to implement this feature because apparently GEM uses this
feature for scrolling. I didn't test myself, but BiFiMSX reported it works now
with this patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12326 7512d703-071c-0410-b5c2-d2bdd895b7d1