Implemented support for MegaDrive joypad
See this feature request:
  [3422738] Please add support for emulating a MegaDrive/Genesis joypad connected through a JoyMega adapter.

It's implemented slightly different than in the feature request though: we
always emulate a 6-button joypad (like a real 6-button joypad, you can force it
in 3-button mode by holding the mode key while plugging in the device). Another
difference is that we allow each host joystick to be used as a MegaDrive
joypad.

Tested with the following tool:
  http://frs.badcoffee.info/files/joymega.lzh
(thanks to mth for doing the actual testing).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12332 7512d703-071c-0410-b5c2-d2bdd895b7d1