Changed the default value for the 'horizontal_stretch' setting from 284 to 280.
The reason for changing this is that since this release the SDL renderer has
support for horizontal_stretch. It can handle arbitrary values for this
setting, but for some specific values (256, 272, 280, 288, 320) we have a
specific implementation that is a bit faster. E.g. for 280 we can stretch 7
input pixels on 8 output pixels.

The old default value for horizontal_stretch was 284, for this value it's hard
to make a specialized stretch routine (one that is significantly faster than
the generic version).

This value 284 was chosen semi-arbitrarily, we picked some value that looked
good. But values close to 284 (like 280 or 288 for which we have specialized
stretch routines) or probably equally good.

So given the choice between 280 and 288, I choose 280 because that stretch
implementation seemed slightly faster (though I didn't test) and because it's
closer to 272, which is the actual number of pixels that the VDP produces per
line.

OpenMSX users who upgrade from 0.8.0 to 0.8.1 will see the following:
- If they used the SDL renderer, the horizontal_stretch effect will be applied
  now (it's as-if the old horizontal_stretch setting was always 320 (=no
  stretch)). So they will notice a big difference, but they will for whatever
  default value we pick (280, 284 or 288).
- If they used the SDLGL-PP renderer and they tweaked the horizontal_stretch
  setting to their liking, then nothing will change: openMSX-0.8.1 will
  continue using their tweaked setting. If they never changed the setting, then
  they will now have a different value. Though, unless you see both values next
  to each other, I doubt many people will notice the difference.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12342 7512d703-071c-0410-b5c2-d2bdd895b7d1