Removed last user of PluggingController::unregisterPluggable()
For some reason (probably historical) CassettePort kept ownership of the
CassettePlayer pluggable itself. PluggingController also wants ownership
of all registered Pluggables, so to avoid a double-free error, an
unregister method was needed.

This patch simplifies the code by simply passing ownership to
PluggingController. It's mostly a preparation for the next patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12386 7512d703-071c-0410-b5c2-d2bdd895b7d1