Fixed serialization version number of the YMF262 class
The serialization (savestate/loadstate) version number for the YMF262 class was
implemented wrongly: the version was set in the .cc file, but it should be set
in the .hh file (so that the version is visible for external code).

The effect was that the class was saved using the layout of version 2 (because
the version check is optimized away for the saving code, we always assume we're
saving using the latest version), but without adding the 'version=2' attribute
in XML. So the loading code sees a class without version attribute (= version
1) and ignores the extra data that was added in version 2.

Luckily this bug doesn't cause problems when loading these 'wrong' savestates:
the only extra field introduced in version 2 was 'alreadySignaledNEW2', and
it's rarely a problem when that field is not loaded correctly. There was no
data removed from version 1 to 2, and all other fields keep their original
meaning.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12395 7512d703-071c-0410-b5c2-d2bdd895b7d1