'Pimpl'-idiom related cleanups.
For some background on the pimpl-idiom see for example here:
  http://herbsutter.com/gotw/_100/

- The more common abbreviation for 'pointer-to-implementation' seems to be
  'pimpl' (we used 'pimple').
- It's possible to better hide the implementation class by making it a
  (private) inner class of the visible class. (As a side effect this forces you
  to think better about the interaction with other (invisible) helper classes
  and this uncovered some small bugs in the Y8950 debug code, see below).

It's possible to do more cleanups in the MSXMotherBoard class (e.g. get rid of
the back-pointer in the MSXMotherBoard::Impl class), but that's something for a
later patch.

This patch also fixes a minor bug in reading the Y8950 registers directly via
the 'MSX Audio regs' debuggable (it went fine on peeking via the IO ports): it
always returned the last written value instead of the value that would actually
be read with an IN instruction (can make a difference for some ADPCM and the
keyboard-input registers).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12397 7512d703-071c-0410-b5c2-d2bdd895b7d1