Added SectorBuffer union
This allows to remove the (dangerous) reinterpret_cast's between the different
interpretations of sector data (e.g. between a raw byte array of 512 bytes and
an array of 16 directory entries).

This patch (mainly) use SectorBuffer internally in DirAsDSK and MSXtar. The next
patch will use it more intensively.

This patch adds some (slightly ugly) '.raw' field selectors on the uses of the
new SectorBuffer objects. Many of those can be removed again in the next patch.