In VDPCmdEngine replace Clock with EmuTime
We basically only used a Clock object to convert a number of VDP clock cycles
(stored as an integer into a table) into an EmuDuration and add that duration
(in the command engine loop) to the current time. This conversion requires a
multiplication (at run-time). But we can as well immediately store the
EmuDuration in the table (so do the multiplication at compile time).

In the (old) code, that conversion happened inside the command engine inner
loop(s), though gcc was already able to move it outside. So this patch doesn't
measurably improve performance. But still the resulting code is a couple of
kB's smaller.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12401 7512d703-071c-0410-b5c2-d2bdd895b7d1