Simplify calculation of future moments in time
If you have the current EmuTime and want to know the EmuTime for let's say
15ms in the future. Before this patch, we typically had to do something like:

  Clock<1000> clock(now);        // 1kHz clock , initialized with current time
  clock += 15;                   // advance 15ms
  someFunction(clock.getTime()); // use that timestamp

This patch allows to write it also like this:

  someFunction(now + EmuDuration::msec(15));

That's much simpler code (it also more directly expresses what you want).
Though gcc was already able to optimize the old (more complex) code to the
same instruction sequence as the new (simpler) code.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12402 7512d703-071c-0410-b5c2-d2bdd895b7d1