Refactored the WD2793 datatransfer mechanism
More specifically refactored the logic of when the DTRQ signal should be
active. The old implementation was a hack: e.g. for a sector read command,
after the data-register was read, the DTRQ signal went down for a fixed amount
of time. This duration was tuned so that it worked with a Philips disk ROM (and
it turned out that for most other disk ROMs this hack was also good enough). In
reality the DTRQ signal should become active again when the next byte is
available (so NOT after a FIXED duration).

The write sector and write track commands had similar timing hacks. The format
routine in the disk ROM of the Spectravideo_SVI-738 machine is implemented
slightly different than the routine in a Philips machine: it is slightly faster
(takes less Z80 cycles per iteration). So because of this timing hack, during
one rotation of the floppy disk, it was able to write slightly more bytes, and
this was enough to cause a buffer overflow (resulting in an assert).

In this patch we (de)activate the DTRQ signal at the proper time. We also
implement the LOST_DATA status bit (set when the CPU didn't read/write the
data register before the next byte has arrived/is needed, no 'correct' MSX
software should ever trigger this).

The new implementation is also simpler than the old one (mainly because it has
less internal state, and thus less invariants that should be kept).

Because the old and the new timing are slightly different, it is possible that
old replays don't sync anymore. Though I believe (I didn't test) that when there's
no disk inserted, the timing is still exactly the same, so replays for ROM games
should still be fine.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12407 7512d703-071c-0410-b5c2-d2bdd895b7d1