Fixed Y8950::peekReg()
Manuel noticed the 'vu_meters' script was broken for MSX-AUDIO. After some
debugging it turned out the problem was that 'debug read "MSX-Audio regs" ...'
(almost) always returned the value 255. This change was introduced in rev12397,
where I 'fixed' this debuggable to not directly return the value of the
internal register file, but go via the peekReg() method which also takes
'behavior' into account (e.g. reading the adpcm data register returns a value
that is changing in time). Though it turned out that there was also a bug in
the peekReg() method itself, this method only worked for the registers with
extra behavior. Fixed now so that both ways work for both type of registers.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12408 7512d703-071c-0410-b5c2-d2bdd895b7d1