Removed limitation for having aligned MSX memory devices
Before this patch, a memory mapped MSX device always needed to start and stop
at a memory address that was a multiple of 0x100 bytes. The reason for this was
the internal memory caching system (purely an emulator implementation aspect, a
real MSX doesn't have such a system). If a memory region doesn't start or stop
at a multiple of 0x100 bytes, it only spans a partial cacheline. And partial
cachelines are not supported (it also won't be efficient).

This patch removes this alignment constraint by making the unaligned begin and
end ranges of the device uncachable. There's a tiny runtime overhead for
checking the alignment, but that only happens on querying a cacheline, not on
each memory access, so this shouldn't be a problem. (Even more, you only pay
for this when there are multiple devices in the same 16kB page).

Up till now this alignment constraint wasn't a problem in any of the msx
machines or extensions. But Manuel is working on a machine config (HC95??)
where it is convenient to specify unaligned memory regions (both for partial
mirroring of a device in some other slot and for actual device grouping in a
single slot).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12419 7512d703-071c-0410-b5c2-d2bdd895b7d1