Fixed bug in floppy-drive index pulse generation.
When there was no disk inserted in the drive and the motor was spinning, before
this patch, openMSX did still generate index pulses. I tested on a real NMS8250
machine that index pulses are only generated when then the motor is spinning
AND there's disk inserted.

When looking at the code, this behaviour was probably the intended behavior,
but there were some missing parenthesis.

We found out there was 'some' problem when the newly implemented 'VictorFDC'
gave the message 'File not found' on a FILES command in MSX-BASIC when executed
with no disk inserted. 'zeilemaker54' gave some valuable hints to locate the
actual bug. See his post of 24-02-2012, 23:55:
  http://www.msx.org/forum/msx-talk/hardware/sanyo-mfd-001-fdc?page=2
After this fix, the above command correctly gives a 'Disk offline' error
message.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12423 7512d703-071c-0410-b5c2-d2bdd895b7d1