Added motor_off_timeout_ms disk drive config parameter
Added this to be able to emulate the delayed motor off behavior of some Sony
MSX machines. In reality this behavior is implemented in the CXD1032 controller
chip. But in openMSX I've implemented it in the DiskDrive code. The advantage
is that it is now configurable for all MSX machines (not only those that have
this controller chip). Though you could also see it as a disadvantage: now when
creating a machine configuration you have to specify the timeout value instead
of specifying it uses a certain controller chip.

There was also a more technical (and openMSX specific) reason to implement this
in the DiskDrive code rather than the 'FDC mapping' code: We use some
heuristics to guess when an MSX program is loading or saving from/to disk (to
be able to do 'fullspeedwhenloading'). Currently those heuristics are hooked
into the DiskDrive code (it looks at the motor status and head movements). If
the motor delay was implemented external to the DiskDrive code, the heuristics
would break: we would assume the MSX program was still loading/saving for the
duration of the motor timeout phase.

Note that this patch doesn't actually use the new config parameter yet.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12425 7512d703-071c-0410-b5c2-d2bdd895b7d1