Fixed some error/warnings reported by cppcheck
Cppcheck is a static analysis tool for c++. When you run it on the openMSX
source tree, it reports quite some error/warnings. Though most seem to be
false-positives or harmless-and-not-easy-to-fix. I didn't yet go over the full
list in detail, but in this commit are already a few fixes for items reported
by cppcheck.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12427 7512d703-071c-0410-b5c2-d2bdd895b7d1