Fixed assert in SSE/MMX version of some LineScaler routines:
To reproduce:
  - start gfx9000 version of symbos
  - in the configuration screen switch to 768x240

The crash occured in the SSE version of the scale_2on1 routine. That routine
assumed that it could always work on a multiple of 4 pixels (so that the
MMX/SSE instructions can do aligned/full-word accesses). Usually this is the
case except for gfx9000-overscan modes in combination with the sdlgl-pp
renderer.

Fixed by extending the asm routines to do the initial part in asm and the last
few pixels with the slower c++ version of the routine.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12437 7512d703-071c-0410-b5c2-d2bdd895b7d1