Various WD2793 improvements related to low level disk access
The most important change in this patch is to convert WD2793 from the old
access methods in Disk/DiskDrive to the new ones (see previous patch).

This change in interface allows to properly access disks with non-standard
disk formats (e.g. disks with some type of copy-protection). Though at the
moment we don't have support yet for a disk image format that can actually
represent such disks. This will be changed in one of the follow up patches.

Other changes in this patch (made possible by the interface change) are:
 - implementation of the "read track" command
 - implementation of the "read address" command
 - moved the guts of the "write track" command implementation from
   the Disk class to the WD2793 class where it belongs (the Disk class
   doesn't need to know about the special WD2793 format codes).

Other changes made possible, but not yet (fully) implemented are:
 - some details about CRC errors in address mark
 - handling of 'deleted' sectors
 - handling of sector sizes other than 512 bytes (mostly done, but need
   to check some details)
 - verification of the track and head information stored in the track

This last class of changes is left as TODO's for the future (I don't yet have
follow up patches). Though for many MSX disk protections the current
implementation is already good enough. Especially if the protection works on
all MSX machines (meaning it doesn't rely on a specific FDC).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12443 7512d703-071c-0410-b5c2-d2bdd895b7d1