Removed old low-level Disk/DiskDrive access interface
As mentioned in the previous patches: the WD2793 and TC8566AF controllers
have been converted from the old to the new Disk interface, so now the old
interface can be removed.

Note that many of the methods in the old interface were never used (e.g.
readTrackData()) or even never implemented (getTrackHeader(),
getSectorHeader()) or implemented but at the wrong place
(SectorBasedDisk::writeTrackDataImpl() that knew about WD2793 commands).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12445 7512d703-071c-0410-b5c2-d2bdd895b7d1