Improved accuracy of disk drive rotation
Before this patch, when the drive motor was stopped and later started again,
the orientation (=rotation angle) of the disk was reset to zero degrees. After
this patch, we continue rotating from the angle we stopped at.

This change was needed to make the 'Sunrise disk copy-protection'-test pass on
a TurboR machine (tested with a DMK image of 'Shrines of Enigma'). This disk
has a track which contains a sector with the same ID multiple times. It reads
this sector several times and expects to see different results. Though without
this disk drive improvement it turned out we always read the same version and
thus the test failed. On a Philips MSX2 machine the drive motor is not turned
off for the duration of the copy-protection-test, and so this change doesn't
have any effect (the test already passed).

Note that a real disk drive has some inertia (both for spinning up and spinning
down). This is currently not emulated. Also need to investigate how a real
TurboR behaves when the motor is turned off for a short amount of time. Does it
really stop rotating (not taking inertia into account) or does it have some
hardware timeout mechanism?

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12448 7512d703-071c-0410-b5c2-d2bdd895b7d1