Fixed behavior of sectors with CRC error on TC8566AF
When reading a sector that has a CRC error (from a DMK disk image) we did set
the status bit that indicates there was a CRC error, But we forgot to set the
bit that indicates 'abnormal command termination'. It seems the turboR disk rom
first checks this general status bit before looking at the specific bits,
because the effect was that this sector with CRC error was read correctly
(while on a WD2793 reading the same sector did result in a CRC error).

This patch also sets the 'abnormal termination' bit on 'write protected'
errors. Even though it already worked before this patch. We also forgot to set
this bit on 'overrun' errors.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12463 7512d703-071c-0410-b5c2-d2bdd895b7d1