Fixed assert in ResampleHQ
The HQ resampler has an internal buffer that holds input samples (= samples
from the MSX sound devices). In the following situation this buffer was too
small and an assert was triggered:
 - A sound device with a high native sample frequency. E.g. the PSG has a
   native frequency of 224kHz. With a higher native frequency, you need more
   input samples to generate the same amount of output samples.
 - Setting a lower value for the 'frequency' setting. E.g. 'set frequency 22050'.
 - Setting a higher value of the 'samples' setting. E.g. 'set samples 8192'.
   The last two items cause openmsx to generate chunks of host samples at a
   lower rate (either because we require less host samples per time or because
   we want larger chunks at a time). The input frequency remains the same so,
   for each chunk there are now more input samples.
To actually trigger the assert, you needed a combination of these 3 items.

Fixed by dynamically growing the buffer when we notice it's too small. An
alternative fix was to increase the buffer size so that it's always big enough.
Though that would waste memory for the majority of the cases where the current
buffer size was already big enough.

Another alternative fix is to keep a (smaller) fixed-size buffer, but to
process the samples in multiple smaller chunks. This may have the advantage
that there's a higher chance that the data fits in the CPU's data cache, but it
also adds more overhead. I didn't (yet) implement this because it's more
complex and it's not obvious it will actually increase performance. And the HQ
resampler should anyway not be used on less powerful hosts (like Dingoo).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12464 7512d703-071c-0410-b5c2-d2bdd895b7d1