Fixed assert in SoundDevice
After the previous patch I could still trigger an assert when I set the
'frequency' and the 'samples' settings combined to extreme values. E.g.
  set frequency 11025
  set samples 8192

The assert failed because the resample factor for the PSG '224kHz -> 11kHz' was
bigger than the maximal assumed factor of '16'.

The code needed this maximal resample factor to allocate a buffer big enough
to hold all the (native) samples of all the individual channels of one sound
device. So it allocated a buffer that can hold:
 - 24 stereo channels (OPL4 has 24 stereo channels)
 - maximal 8192 host samples multiplied by a maximal resample factor of 16 (but
   that value 16 was too small) (only PSG and SCC can have such a high resample
   factor)
In practice this buffer was too large because the combination of 24 stereo
channels and an extreme high resample factor cannot occur simultaneously. So
the assert was actually harmless.

Also this buffer is not that frequently used. It's only used when the data for
an individual sound channel is needed separately, so when this channel is being
recorded or is muted or if the channels have a non-center balance setting. The
current buffer size was about 24MB(!) and typically nothing of this buffer was
actually used. Though this is not as bad as it sounds: on a modern OS (I
verified on Linux) with a big global array you only 'pay' for the memory that
you actually use (of course the cost of the address space remains, but there's
usually plenty of address space).

To fix this assert, I now dynamically grow the buffer when it's too small. The
initial buffer size is zero and in a typical openmsx session (e.g. no channel
recording) the buffer size will stay zero. (After this patch the virtual memory
size of the openmsx process has shrunk by 24MB, though actual used memory
remained the same).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12465 7512d703-071c-0410-b5c2-d2bdd895b7d1