Added support for non-fixed track lengths
Normally a track on a 3.5" double-density floppy disk contains 6250 bytes:
  datarate: 250kbps
  rotation speed: 300rpm
Though on a real diskdrive the rotation speed is not exactly 300rpm (and maybe
the datarate is also not exactly 250kpbs). So the track does also not contain
exactly 6250 bytes. E.g. one disk I just tested has a track length of only 6219
bytes (tested by counting the number of bytes returned by the WD2793 read track
command).

It's also the case that the different tracks on the same disk can (slightly)
differ in length. E.g. for the same disk I tested above the track length varied
between 6213..6229. Though unfortunately the DMK disk image format does not
support different lengths per track (it only allows to specify one track length
for the whole disk).

This patch extends the FDC code to take the actual track length into account
(this changes the timing of the block transfer commands). It also changes the
DMKDiskImage class to not anymore truncate/extend the tracks from the diskimage
to exactly 6250 bytes.

The motivation for adding this it to make it easier to automatically generate
DMK files from real disks. E.g. for the disk mentioned above were the track
length varied between 6213 and 6229 it's easier to convert all tracks to e.g.
a length of 6221 than to convert them all to 6250.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12475 7512d703-071c-0410-b5c2-d2bdd895b7d1