Fixed crash in Scale_1on1::operator() on ARM
This regression was introduced in revision 12437 (thanks Juha for bisecting the
problem!). After this revision we still execute some c++ code after an ARM
inline assembly block (before we immediately returned after this block). The
bug was that the constraints on the asm block were wrong. So the compiler
assumed the asm block preserved the content of certain registers, while in
reality the asm code did actually change these registers. So this was already a
latent bug before revision 12437.

Fixed by properly specifying the constaints on the asm block. I now also take
advantage of the fact that the ARM asm code already updates the in- and out-
pointers, so the code is slightly faster now. I manually inspected the
generated code (gcc-4.6.1) and it seems ok. Juha also tested this code on
maemo5 and it seems to work fine.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12479 7512d703-071c-0410-b5c2-d2bdd895b7d1