[1/14] fix inline asm: ResampleHQ.cc
In revision 12479 we fixed a regression. That regression first appeared in
r12437, though actually the code was already wrong before: r12437 only
uncovered a latent bug. The problem was in the constraint specification of an
inline asm block. The constraints failed to mention that some registers changed
value after the asm block. So when in revision 12479 we added some extra code
after the asm block, the compiler has happy to reuse a value that it thought
was still in a register (while in reality that value was changed by the asm
block).

While working on 12479 I noticed there were more asm blocks that had errors in
the constraint specification. Though currently none of these errors actually
cause the compiler to generate wrong code. Still it also are latent bugs. In
this patch series I'm fixing these constraints.

The most common error is (like already explained in the 1st paragraph) failing
to mention that a register is changing. We also often forgot to mention that a
asm block writes to memory. And while I was editing the asm routines I also
switched from positional constraints to named constraints (this actually made
the refactoring easier, because now I was free to reorder the constraints
without changing the meaning of the code).

I tried to be careful while making these changes (but asm is always tricky).
Most changes are independent, so I split this patch in 14 sub-patches. This
should make it easier to bisect a problem, in case I did make a mistake
somewhere.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12482 7512d703-071c-0410-b5c2-d2bdd895b7d1