Don't construct empty strings via string::string("")
Constructing empty std::strings via the default constructor is more efficient
than via the constructor that takes a 'const char*' and pass a "" literal. E.g.
compare these two functions:
   string f1() { return ""; }
   string f2() { return string(); }
The generated code for the former is 32 bytes, 10 instructions (of which one is
a call instruction (so dynamically it are still many more instructions)), the
latter generates 11 bytes, 3 instructions (no call).

Though one major disadvantage is that the 2nd variant is longer to write (and
arguably less clear to read). But in c++11 there's a shorthand notation to
construct objects in a context where the type is known. This allows to write
   string f3() { return {}; }
This is equally short as the 1st variant and means exactly the same as the 2nd
variant (and thus generates identical code).

See patch for more examples of where this construct can be used. In the locations
where "" could not be replaced by {}, I replaced it with 'string()'.

This patch reduces the total code size by about 12kB.